+++ packages/schemas/src/types/logto-config/jwt-customizer.ts
@@ -1,4 +1,4 @@
-import { LogtoJwtTokenKeyType, jwtCustomizerGuard } from '@logto/core-kit/custom-jwt';
+import { jsonObjectGuard } from '@logto/connector-kit';
 import { z } from 'zod';
 
 import { Organizations, Roles, UserSsoIdentities } from '../../db-entries/index.js';
@@ -8,11 +8,16 @@ import { userInfoGuard } from '../user.js';
 
 import { accessTokenPayloadGuard, clientCredentialsPayloadGuard } from './oidc-provider.js';
 
-export {
-  LogtoJwtTokenKeyType,
-  jwtCustomizerGuard,
-  type CustomJwtFetcher,
-} from '@logto/core-kit/custom-jwt';
+export const jwtCustomizerGuard = z.object({
+  script: z.string(),
+  environmentVariables: z.record(z.string()).optional(),
+  contextSample: jsonObjectGuard.optional(),
+});
+
+export enum LogtoJwtTokenKeyType {
+  AccessToken = 'access-token',
+  ClientCredentials = 'client-credentials',
+}
 
 export const jwtCustomizerUserContextGuard = userInfoGuard.extend({
   ssoIdentities: UserSsoIdentities.guard
@@ -80,3 +85,27 @@ export const jwtCustomizerTestRequestBodyGuard = z.discriminatedUnion('tokenType
 ]);
 
 export type JwtCustomizerTestRequestBody = z.infer<typeof jwtCustomizerTestRequestBodyGuard>;
+
+/**
+ * This guard is for cloud API use (request body guard).
+ * Since the cloud API will be use by both testing and production, should keep the fields as general as possible.
+ * The response guard for the cloud API is `jsonObjectGuard` since it extends the `token` with extra claims.
+ */
+const commonJwtCustomizerGuard = jwtCustomizerGuard
+  .pick({ script: true, environmentVariables: true })
+  .required({ script: true })
+  .extend({
+    token: jsonObjectGuard,
+  });
+
+export const customJwtFetcherGuard = z.discriminatedUnion('tokenType', [
+  commonJwtCustomizerGuard.extend({
+    tokenType: z.literal(LogtoJwtTokenKeyType.AccessToken),
+    context: jsonObjectGuard,
+  }),
+  commonJwtCustomizerGuard.extend({
+    tokenType: z.literal(LogtoJwtTokenKeyType.ClientCredentials),
+  }),
+]);
+
+export type CustomJwtFetcher = z.infer<typeof customJwtFetcherGuard>;
