+++ packages/toolkit/core-kit/src/custom-jwt/index.ts
@@ -1,3 +1,2 @@
 export * from './script-execution.js';
 export * from './error-handling.js';
-export * from './types.js';
